--- a/.github/workflows/gradle-build.yml
+++ b/.github/workflows/gradle-build.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2.3.4
-      - uses: gradle/wrapper-validation-action@v1.0.3
+      - uses: gradle/wrapper-validation-action@v1.0.4
   build:
     name: "Build"
     runs-on: ubuntu-latest