--- a/.github/workflows/gradle-build.yml
+++ b/.github/workflows/gradle-build.yml
@@ -26,7 +26,7 @@ jobs:
           java-version: '8'
           distribution: 'adopt'
       - name: Cache SonarCloud packages
-        uses: actions/cache@v2.1.5
+        uses: actions/cache@v2.1.6
         if: env.SONAR_TOKEN != null && env.SONAR_TOKEN != ''
         env:
           SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}