--- a/.github/workflows/gradle-build.yml
+++ b/.github/workflows/gradle-build.yml
@@ -47,6 +47,7 @@ jobs:
         with:
           files: '**/test-results/**/*.xml'
       - name: Analyze with SonarCloud
+        continue-on-error: true
         if: env.SONAR_TOKEN != null && env.SONAR_TOKEN != ''
         uses: eskatos/gradle-command-action@v1
         env: