--- a/.github/workflows/gradle-build.yml
+++ b/.github/workflows/gradle-build.yml
@@ -14,7 +14,7 @@ jobs:
       - uses: actions/checkout@v2.3.4
       - uses: gradle/wrapper-validation-action@v1.0.4
   build:
-    name: "Build"
+    name: "Build and publish"
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2