--- a/build.gradle
+++ b/build.gradle
@@ -33,6 +33,7 @@ nexusPublishing {
         }
     }
 }
+
 configure(project.coreProjects) {
     apply plugin: 'java'
     apply plugin: 'java-library'