--- a/resilience4j-reactor/src/test/java/io/github/resilience4j/reactor/ratelimiter/operator/FluxRateLimiterTest.java
+++ b/resilience4j-reactor/src/test/java/io/github/resilience4j/reactor/ratelimiter/operator/FluxRateLimiterTest.java
@@ -62,7 +62,6 @@ public class FluxRateLimiterTest {
 
         StepVerifier.create(
             Flux.error(new IOException("BAM!"))
-                .log()
                 .transformDeferred(RateLimiterOperator.of(rateLimiter)))
             .expectSubscription()
             .expectError(IOException.class)