--- a/resilience4j-retrofit/build.gradle
+++ b/resilience4j-retrofit/build.gradle
@@ -2,6 +2,7 @@ dependencies {
     compileOnly(libraries.retrofit)
     implementation project(':resilience4j-circuitbreaker')
     implementation project(':resilience4j-ratelimiter')
+    implementation project(':resilience4j-retry')
     testImplementation(libraries.retrofit_test)
     testImplementation(libraries.retrofit_wiremock)
     testImplementation(libraries.retrofit)